WeatherController_GetTemp_NoZipCode_Returns400 is passing
