+++ source/api.tests/WeatherControllerTests.cs
@@ -1,4 +1,8 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Api.Weather;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Moq;
+using System.Threading.Tasks;
 
 namespace Api.Tests
 {
@@ -6,10 +10,30 @@ namespace Api.Tests
     [TestCategory(TestCategories.WeatherAPI)]
     public class WeatherControllerTests
     {
-        [TestMethod]
-        public void WeatherController()
+
+        private (WeatherController weatherController, Mock<IWeatherProvider> weatherProvider) Factory()
         {
-            Assert.IsFalse(true);
+            var weatherProvider = new Mock<IWeatherProvider>();
+            return (new WeatherController(weatherProvider.Object), weatherProvider);
         }
+
+        [TestMethod]
+        public void WeatherController_GetTemp_NoZipCode_Returns400()
+        {
+            // Arrange
+            var returnedMocks = Factory();
+
+            // Act
+            var result = returnedMocks.weatherController.GetTempForZip(0);
+
+            // Assert
+            Assert.AreEqual(400, (result.Result as BadRequestObjectResult).StatusCode);
+        }
+
+        /**
+         * Given an API call
+        * When asking for current temp
+        * Then it calls the weather Api with the correct zip code 
+        */
     }
 }
