+++ source/api/Weather/IWeatherProvider.cs
@@ -1,0 +1,6 @@
+namespace Api.Weather
+{
+    public interface IWeatherProvider
+    {
+    }
+}
