+++ source/api/Weather/WeatherProvider.cs
@@ -1,0 +1,9 @@
+namespace Api.Weather
+{
+    public class WeatherProvider : IWeatherProvider
+    {
+        public WeatherProvider()
+        {
+        }
+    }
+}
