+++ source/web.tests/UnitTest1.cs
@@ -5,9 +5,5 @@ namespace web.tests
     [TestClass]
     public class UnitTest1
     {
-        [TestMethod]
-        public void TestMethod1()
-        {
-        }
     }
 }
