--- source/web/Controllers/WeatherController.cs
@@ -1,44 +1,0 @@
-using System.Linq;
-using Microsoft.AspNetCore.Mvc;
-using web.Providers;
-
-namespace web.Controllers
-{
-    [Route("api/[controller]")]
-    public class WeatherController : Controller
-    {
-        private readonly IWeatherProvider weatherProvider;
-
-        public WeatherController(IWeatherProvider weatherProvider)
-        {
-            this.weatherProvider = weatherProvider;
-        }
-
-        [HttpGet("[action]")]
-        public IActionResult Forecasts([FromQuery(Name = "from")] int from = 0, [FromQuery(Name = "to")] int to = 4)
-        {
-            //System.Threading.Thread.Sleep(500); // Fake latency
-            var quantity = to - from;
-
-            // We should also avoid going too far in the list.
-            if (quantity <= 0)
-            {
-                return BadRequest("You cannot have the 'to' parameter higher than 'from' parameter.");
-            }
-
-            if (from < 0)
-            {
-                return BadRequest("You cannot go in the negative with the 'from' parameter");
-            }
-
-            var allForecasts = weatherProvider.GetForecasts();
-            var result = new
-            {
-                Total = allForecasts.Count,
-                Forecasts = allForecasts.Skip(from).Take(quantity).ToArray()
-            };
-
-            return Ok(result);
-        }
-    }
-}
