--- source/web/Models/WeatherForecast.cs
@@ -1,13 +1,0 @@
-namespace web.Models
-{
-    public class WeatherForecast
-    {
-        public string DateFormatted { get; set; }        
-
-        public string Summary { get; set; }
-
-        public int TemperatureC { get; set; }
-
-        public int TemperatureF => 32 + (int)(TemperatureC / 0.5556);
-    }
-}
