--- source/web/Providers/IWeatherProvider.cs
@@ -1,10 +1,0 @@
-using System.Collections.Generic;
-using web.Models;
-
-namespace web.Providers
-{
-    public interface IWeatherProvider
-    {
-        List<WeatherForecast> GetForecasts();
-    }
-}
