+++ source/web/Startup.cs
@@ -22,9 +22,7 @@ namespace web
             // Add framework services.
             services.AddMvc()
                 .SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-
-            // Simple example with dependency injection for a data provider.
-            services.AddSingleton<Providers.IWeatherProvider, Providers.WeatherProviderFake>();
+            
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
